+++ svs/tests/factories.py
@@ -232,9 +232,9 @@ class FormDataFactory:
 
     # gene lists / regions
 
-    gene_blacklist: typing.List[str] = attr.Factory(list)
+    gene_blocklist: typing.List[str] = attr.Factory(list)
     genomic_region: typing.List[str] = attr.Factory(list)
-    gene_whitelist: typing.List[str] = attr.Factory(list)
+    gene_allowlist: typing.List[str] = attr.Factory(list)
 
     # tads