+++ variants/queries.py
@@ -984,10 +984,10 @@ class ExtendQueryPartsTranscriptCodingFilter(ExtendQueryPartsBase):
 class ExtendQueryPartsGeneListsFilter(ExtendQueryPartsBase):
     def extend_conditions(self, _query_parts):
         result = []
-        if self.kwargs["gene_blacklist"]:
-            result.append(not_(self._build_list(self.kwargs["gene_blacklist"])))
-        if self.kwargs["gene_whitelist"]:
-            result.append(self._build_list(self.kwargs["gene_whitelist"]))
+        if self.kwargs["gene_blocklist"]:
+            result.append(not_(self._build_list(self.kwargs["gene_blocklist"])))
+        if self.kwargs["gene_allowlist"]:
+            result.append(self._build_list(self.kwargs["gene_allowlist"]))
         return result
 
     def _build_list(self, gene_list):